ml-sidecar/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
essentia-tensorflow==2.1b6.dev1110
librosa==0.10.2
madmom==0.16.1
torch==2.2.2
onnxruntime-gpu==1.17.1
transformers==4.40.0
demucs==4.0.1
pyacoustid==1.3.0
mutagen==1.47.0
soundfile==0.12.1
numpy==1.26.4
scipy==1.13.0
pydantic==2.7.3
python-multipart==0.0.9
